metro_assist/requirements.txt
Django==5.0.6
psycopg2_binary==2.9.9
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pytz==2024.1